docs: improve responsive behaviour of demo app
