+++ apps/demo/config/blocks/Stats/styles.module.css
@@ -10,13 +10,22 @@
   grid-gap: 72px;
   align-items: center;
   justify-content: space-between;
-  padding: 64px;
+  margin: 0 auto;
+  max-width: 768px;
+  padding: 64px 16px;
 }
 
 @media (min-width: 768px) {
   .Stats-items {
+    padding: 64px 24px;
+  }
+}
+
+@media (min-width: 1024px) {
+  .Stats-items {
     grid-template-columns: 1fr 1fr;
-    padding: 128px;
+    padding: 128px 24px;
+    max-width: 100%;
   }
 }
