+++ apps/demo/config/components/Header/styles.module.css
@@ -1,0 +1,33 @@
+.Header {
+  align-items: center;
+  display: flex;
+  margin-left: auto;
+  margin-right: auto;
+  max-width: 1280px;
+  padding: 24px 16px;
+}
+
+@media (min-width: 768px) {
+  .Header {
+    padding: 24px;
+  }
+}
+
+.Header-logo {
+  font-size: 24px;
+  font-weight: 800;
+  letter-spacing: 1.4;
+  opacity: 0.8;
+}
+
+.Header-items {
+  display: flex;
+  gap: 24px;
+  margin-left: auto;
+}
+
+@media (min-width: 1024px) {
+  .Header-items  {
+    gap: 32px;
+  }
+}
