+++ apps/demo/config/components/Section/styles.module.css
@@ -1,6 +1,13 @@
 .Section:not(.Section .Section) {
-  padding-left: 24px;
-  padding-right: 24px;
+  padding-left: 16px;
+  padding-right: 16px;
+}
+
+@media (min-width: 768px) {
+  .Section:not(.Section .Section) {
+    padding-left: 24px;
+    padding-right: 24px;
+  }
 }
 
 .Section:not(.Section .Section) .Section-inner {
